Formatting rearranging and cleaning
Changed order of plots to match analysis, changed order of plots to match the pairplot formatting (e.g. age --> current rank --> best rank).
Moved tennis_split to later in the code when it is needed to be used and removed tennis_split_2 since we are only using the log_prize_money version.